优化

+++ app/src/main/java/com/chengmao/mapdemo/track/EndTrackActivity.java
@@ -54,7 +54,7 @@ public class EndTrackActivity extends BaseActivity {
     private int mType = 1;
     private String name;
     private EditText et_desc;
-    private float space;
+    private int space;
     private long time;
 
     @Override
@@ -68,7 +68,7 @@ public class EndTrackActivity extends BaseActivity {
         start_location = (LocationBean) intent.getSerializableExtra("start_location");
         end_location = (LocationBean) intent.getSerializableExtra("end_location");
         startTrackBean = (StartTrackBean) intent.getSerializableExtra("track_bean");
-        space = intent.getFloatExtra("space", 0);
+        space = intent.getIntExtra("space", 0);
         time = intent.getLongExtra("time", 0);
         et_name = findViewById(R.id.et_name);
         tv_type = findViewById(R.id.tv_type);
